chore: rename bar class to Bar, rename types.hpp to Bar.hpp
